% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFrench.R
\name{extractFrench}
\alias{extractFrench}
\title{extractFrench}
\usage{
extractFrench(path_out = "./data")
}
\arguments{
\item{path_out}{path in which CSV should be unpacked}
}
\value{
Fama/French 3 Factors daily data
}
\description{
This function automatically downloads the relevant csv file for risk free
rates from Kenneth R. Frenchs Website (https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html)
}
