% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPolygonMeta.R
\name{extractPolygonMeta}
\alias{extractPolygonMeta}
\title{extractPolygonMeta}
\usage{
extractPolygonMeta(apikey, ticker)
}
\arguments{
\item{apikey}{Personal API key for polygon.io}

\item{ticker}{ticker for underlying}
}
\value{
metadata of underlying firm
}
\description{
extractPolygonMeta
}
