% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPolygonOptContracts.R
\name{extractPolygonOptContracts}
\alias{extractPolygonOptContracts}
\title{extractPolygonOptContracts}
\usage{
extractPolygonOptContracts(
  apikey,
  tickers,
  from,
  to,
  database,
  adjusted = "true",
  limit = 50000,
  limitedAPIcalls = TRUE
)
}
\arguments{
\item{limitedAPIcalls}{}
}
\description{
extractPolygonOptContracts
}
