% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPolygonStock.R
\name{extractPolygonStock}
\alias{extractPolygonStock}
\title{extractPolygonStock}
\usage{
extractPolygonStock(
  apikey,
  limit = 100,
  ticker,
  from,
  to,
  adjusted = "true",
  sort = "asc"
)
}
\arguments{
\item{apikey}{Personal API key for polygon.io}

\item{limit}{Limit the number of results returned (maximum is 50000)}

\item{ticker}{Ticker symbol of the stock/equity}

\item{from}{The start of the aggregate time window. Either a date with the format YYYY-MM-DD or a millisecond timestamp}

\item{to}{The end of the aggregate time window. Either a date with the format YYYY-MM-DD or a millisecond timestamp}

\item{adjusted}{Whether or not the results are adjusted for splits}

\item{sort}{Sort the results by timestamp (asc or desc)}
}
\value{
historical stock data
}
\description{
extractPolygonStock
}
