% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPolygonStocks.R
\name{extractPolygonStocks}
\alias{extractPolygonStocks}
\title{extractPolygonStocks}
\usage{
extractPolygonStocks(
  apikey,
  tickers,
  database,
  from,
  to,
  limitedAPIcalls,
  limit = 50000,
  adjusted = "true"
)
}
\arguments{
\item{apikey}{Personal API key for polygon.io}

\item{tickers}{Limit the number of results returned (maximum is 1000)}

\item{database}{local rsqlite database to write data on}

\item{from}{The start of the time window. Either a date with the format YYYY-MM-DD or a millisecond timestamp.}

\item{to}{The end of the time window. Either a date with the format YYYY-MM-DD or a millisecond timestamp.}

\item{limitedAPIcalls}{set TRUE if free plan (only makes 5 API calls per minute)}

\item{limit}{Limit of number of queries (maximum is 50000)}

\item{adjusted}{Whether or not the results are adjusted for splits ("true" or "false")}
}
\value{
nothing, data is written on database
}
\description{
Get relevant stock data for given ticker(s)
}
