% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPolygonTickers.R
\name{extractPolygonTickers}
\alias{extractPolygonTickers}
\title{extractPolygonTickers}
\usage{
extractPolygonTickers(apikey, search = NA, limit = 100)
}
\arguments{
\item{apikey}{Personal API key for polygon.io}

\item{search}{Search for terms within the ticker and/or company name}

\item{limit}{Limit the number of results returned (maximum is 1000)}
}
\value{
data.frame tickers of Stocks/Equities, Indices, Forex, and Crypto
}
\description{
Function to search for tickers
}
