% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optipod.R
\name{optipod}
\alias{optipod}
\title{optipod}
\usage{
optipod(
  mu = mu,
  K = K,
  r = r,
  ttm = ttm,
  Lweights = Lweights,
  multiplicationFactor = 5,
  visualise = F,
  ticker = NA,
  date = NA
)
}
\arguments{
\item{mu}{vector of option prices}

\item{K}{vector of strike prices}

\item{r}{annual risk free interest rate (in decimals)}

\item{ttm}{time to maturity in years}

\item{Lweights}{weights assigned to option contract}

\item{multiplicationFactor}{multiplication factor to define upper bound of domain}

\item{visualise}{if TRUE, plots the resulting probability density function}

\item{ticker}{ticker which will be displayed in title of plot}

\item{date}{date which will be displayed in title of plot}
}
\value{
data.frame with necessary results (one row)
}
\description{
Function to estimate the option implied probability of default according
to "Vilsmeier, J. (2014). Updating the option implied probability of default
methodology." Available at SSRN 2797025. Note that the code was provided
by Johannes Vilsmeier but was refactored, therefore possible errors might
were included in original code.
}
